- Before the deletion process has started all SMS in the haiku bin will be sorted by date (newer SMS will be higher up in the list)
- When the deletion process has started the user selects where he wants
to place the SMS. He does this by dragging an object (either an SMS, a
conversation or a date) to the bin. This will open the bin and the user
can then place the object by dropping it into the right position. The
object can only be placed between words. If the user drops the object on
a word the object will be placed after that word.

The position the user has chosen is shown with a marker (this is
difficult to see due to the finger blocking the view most of the time).
This is useful, for instance, when the user is hovering over a word to
see that the object will actually be placed after the word.

If there are more than one SMS the SMS will be added at the chosen
position sorted by date (newer SMS will be higher up in the list).

The user can scroll while dragging by placing the finger at the top of
the list (to scroll up) or at the bottom of the list (to scroll down).