- SMS now have rounded corners
- The text area will now scroll automatically when the top row(s) are
removed. If the user is not at the top he will not notice the scrolling.
If he is at the top the words will fall down as usual and when all
animations has ended the ScrollView will scroll so that there will be no
empty rows at the top. The scrolling stops happening when the rows that
are left only cover 70% of the text area (so the words will be
centered).