- The share message now ends with "-Haiku" with offset to the right so that the receiver understands what it is.
- The date view is slightly (10%) smaller now.

- The "Time" and year texts are now slightly further down.

- The deleting process is now slightly faster if there is much text. I
have added constants that can easily be changed if it is not fast
enough/too fast.

- Fixed a bug which caused the full time not to be shown in the SMS.

- The Haiku conversation is now always at the top.

- The App now supports conversations with more than 2 participants. Max
3 names of the conversation is shown in the conversation list, if there
are more, a "+" sign is shown in the bottom right. In the conversation
(SMS view), only the first names (currently 2 (it depends on the size of
the text)) are shown. By pressing the names a bigger scrollable list is
shown which contains all names, pressing it again will close it.

- If the user sent an SMS in a conversation with more than 1 recipient
the App would show the SMS several times (one for each recipient). This
is because they are stored as different SMS in the phone. This is now
fixed and these SMS are now counted as one instead.