- Removed sample contacts
- "..." instead of "+" to show that there are more contacts that do not
fit on the screen

- Removed HaikuProgressBar and let the text take up the space instead

- The dates has been moved in the BinView

- The slider in the BinView now always shows

- Fixed a bug which sometimes caused the user being unable to scroll
after deleting

- Can now drag the views that are in the small bin view

- Complete rework of how the themes work: They can only be added from a
conversation and they now add the SMS from that conversation that fit
the theme instead of filtering the words in the bin. An SMS is added if
it has at least one word which is in the theme's dictionary. Those words
have a higher chance of being used in the final haiku