- You can't add date or theme when the deletion has started
- can no longer add future dates

- The themes and the dates has to both be true. If you add a date and
then a theme it will remove SMS from the bin. If there are both dates
and themes when you add a date or theme SMS will be added. (d1 OR d2 OR
d3...) AND (t1 OR t2 OR t3...) for all SMS in the bin. Each conversation
has its own dates and themes

- If you remove a theme/date only the SMS which were added with that
theme/date will be removed

- Added a maximum SMS cap i the bin. No more than that cap can be added
(currently 60)

- Fixed two crashes